+++ src/actions/types.ts
@@ -6,13 +6,7 @@ export const CLEAR_BOOKMARKS = 'CLEAR_BOOKMARKS';
 
 export const CLEAR_SEARCH_HISTORY = 'CLEAR_SEARCH_HISTORY';
 
-export const OPEN_DETAIL_INIT = 'OPEN_DETAIL_INIT';
-
-export const OPEN_DETAIL_PENDING = 'OPEN_DETAIL_PENDING';
-
-export const OPEN_DETAIL_SUCCESS = 'OPEN_DETAIL_SUCCESS';
-
-export const OPEN_DETAIL_ERROR = 'OPEN_DETAIl_ERROR';
+export const OPEN_DETAIL = 'OPEN_DETAIL';
 
 export const QUERY_SEARCH_INIT = 'QUERY_SEARCH_INIT';
 
@@ -21,6 +15,30 @@ export const QUERY_SEARCH_PENDING = 'QUERY_SEARCH_PENDING';
 export const QUERY_SEARCH_SUCCESS = 'QUERY_SEARCH_SUCCESS';
 
 export const QUERY_SEARCH_ERROR = 'QUERY_SEARCH_ERROR';
+
+export const LOAD_METRIC_INIT = 'LOAD_METRIC_INIT';
+
+export const LOAD_METRIC_PENDING = 'LOAD_METRIC_PENDING';
+
+export const LOAD_METRIC_SUCCESS = 'LOAD_METRIC_SUCCESS';
+
+export const LOAD_METRIC_ERROR = 'LOAD_METRIC_ERROR';
+
+export const LOAD_INDOM_INIT = 'LOAD_INDOM_INIT';
+
+export const LOAD_INDOM_PENDING = 'LOAD_INDOM_PENDING';
+
+export const LOAD_INDOM_SUCCESS = 'LOAD_INDOM_SUCCESS';
+
+export const LOAD_INDOM_ERROR = 'LOAD_INDOM_ERROR';
+
+export const LOAD_METRIC_INDOM_INIT = 'LOAD_METRIC_INDOM_INIT';
+
+export const LOAD_METRIC_INDOM_PENDING = 'LOAD_METRIC_INDOM_PENDING';
+
+export const LOAD_METRIC_INDOM_SUCCESS = 'LOAD_METRIC_INDOM_SUCCESS';
+
+export const LOAD_METRIC_INDOM_ERROR = 'LOAD_METRIC_INDOM_ERROR';
 
 export const CLEAR_RESULTS = 'CLEAR_RESULTS';
 
@@ -71,38 +89,91 @@ export type QuerySearchAction =
   | QuerySearchSuccessAction
   | QuerySearchErrorAction;
 
-export interface OpenDetailInitAction {
-  type: typeof OPEN_DETAIL_INIT;
-}
-
-export interface OpenDetailPendingAction {
-  type: typeof OPEN_DETAIL_PENDING;
-}
-
-export interface OpenDetailSuccessAction {
-  type: typeof OPEN_DETAIL_SUCCESS;
-  payload: EntityData;
-}
-
-export interface OpenDetailErrorAction {
-  type: typeof OPEN_DETAIL_ERROR;
+export interface OpenDetailAction {
+  type: typeof OPEN_DETAIL;
 }
 
 export interface ClearResultsAction {
   type: typeof CLEAR_RESULTS;
 }
 
-export type OpenDetailAction =
-  | OpenDetailInitAction
-  | OpenDetailPendingAction
-  | OpenDetailSuccessAction
-  | OpenDetailErrorAction;
+export interface LoadMetricInitAction {
+  type: typeof LOAD_METRIC_INIT;
+}
+
+export interface LoadMetricPendingAction {
+  type: typeof LOAD_METRIC_PENDING;
+}
+
+export interface LoadMetricSuccessAction {
+  type: typeof LOAD_METRIC_SUCCESS;
+  payload: MetricData;
+}
+
+export interface LoadMetricErrorAction {
+  type: typeof LOAD_METRIC_ERROR;
+}
+
+export type LoadMetricAction =
+  | LoadMetricInitAction
+  | LoadMetricPendingAction
+  | LoadMetricSuccessAction
+  | LoadMetricErrorAction;
+
+export interface LoadMetricIndomInitAction {
+  type: typeof LOAD_METRIC_INDOM_INIT;
+}
+
+export interface LoadMetricIndomPendingAction {
+  type: typeof LOAD_METRIC_INDOM_PENDING;
+}
+
+export interface LoadMetricIndomSuccessAction {
+  type: typeof LOAD_METRIC_INDOM_SUCCESS;
+  payload: IndomData;
+}
+
+export interface LoadMetricIndomErrorAction {
+  type: typeof LOAD_METRIC_INDOM_ERROR;
+}
+
+export type LoadMetricIndomAction =
+  | LoadMetricIndomInitAction
+  | LoadMetricIndomPendingAction
+  | LoadMetricIndomSuccessAction
+  | LoadMetricIndomErrorAction;
+
+export interface LoadIndomInitAction {
+  type: typeof LOAD_INDOM_INIT;
+}
+
+export interface LoadIndomPendingAction {
+  type: typeof LOAD_INDOM_PENDING;
+}
+
+export interface LoadIndomSuccessAction {
+  type: typeof LOAD_INDOM_SUCCESS;
+  payload: IndomData;
+}
+
+export interface LoadIndomErrorAction {
+  type: typeof LOAD_INDOM_ERROR;
+}
+
+export type LoadIndomAction =
+  | LoadIndomInitAction
+  | LoadIndomPendingAction
+  | LoadIndomSuccessAction
+  | LoadIndomErrorAction;
 
 export type SearchAction =
   | ClearBookmarksAction
   | ClearSeachHistoryAction
   | QuerySearchAction
   | OpenDetailAction
+  | LoadIndomAction
+  | LoadMetricAction
+  | LoadMetricIndomAction
   | AddBookmarkAction
   | ClearResultsAction;
 
@@ -146,25 +217,34 @@ export interface SearchResult extends TrackableStatus {
   data: SearchResultData | null;
 }
 
-export interface MetricData {
+export interface MetricDetailState {
   type: EntityType.Metric;
-  metric: PmApiMetricEndpointMetricResponse;
-  indom?: PmApiIndomEndpointResponse;
+  metric: MetricDataState;
+  indom?: IndomDataState;
 }
 
-export interface InstanceData {
+export interface InstanceDetailState {
   type: EntityType.Instance;
 }
 
-export interface InstanceDomainData {
+export interface InstanceDomainDetailState {
   type: EntityType.InstanceDomain;
+  indom: IndomDataState;
 }
 
-export type EntityData = MetricData | InstanceData | InstanceDomainData;
+export type MetricDataState = MetricData & TrackableStatus;
 
-export interface EntityDetail extends TrackableStatus {
-  data: EntityData | null;
+export type IndomDataState = IndomData & TrackableStatus;
+
+export interface MetricData {
+  data: PmApiMetricEndpointMetricResponse | null;
 }
+
+export interface IndomData {
+  data: PmApiIndomEndpointResponse | null;
+}
+
+export type DetailState = MetricDetailState | InstanceDetailState | InstanceDomainDetailState | null;
 
 export interface TrackableStatus {
   status: FetchStatus;
@@ -176,7 +256,7 @@ export interface SearchState {
   history: SearchQuery[];
   result: SearchResult;
   bookmarks: BookmarkItem[];
-  entity: EntityDetail;
+  entity: DetailState;
 }
 
 export enum SearchEntity {